recoveryapi: cleanup and remove
