--- internal/recoveryapi/recoveryapi.go
@@ -1,9 +1,0 @@
-// Copyright 2024 Edgeless Systems GmbH
-// SPDX-License-Identifier: AGPL-3.0-only
-
-package recoveryapi
-
-//go:generate protoc --go_out=. --go_opt=paths=source_relative --go-grpc_out=. --go-grpc_opt=paths=source_relative recoveryapi.proto
-
-// Port is the port of the coordinator API.
-const Port = "1314"
